requirements.txt
streamlit==1.31.1
opencv-python-headless==4.9.0.80
numpy==1.24.4
Pillow==10.2.0
tensorflow==2.15.0
streamlit-webrtc==0.45.0
